feat: add support for lazy refresh
By default the connector will continue to use a refresh ahead strategy,
where client certificates are refreshed by a background thread. The
lazy strategy by comparison is useful for when the Connector runs in
serverless environments and background threads may not run reliably,
e.g., when the CPU is throttled.

This commit adds the lazy refresh strategy and updates the configuration
documentation to demonstrate how to use the feature.

Fixes #565