chore: replace AlloyDBFeature with static JSON config
This commit removes the references to the GraalVM SDK and also removes
the AlloyDBFeature class. Instead, the commit uses JSON to configure the
native image.

This resolves a few problems:
- The code wasn't consuming the latest GraalVM SDK from the parent POM
- The presence of newer GraalVM breaks tests on Java 8.
- The native-image META-INFO directory wasn't matching
  groupId/artifactId for the Java Connector.

Now, with this commit, there is no need to reference GraalVM SDK at all.
And there is no conflict with Java 8 tests.