chore: replace AlloyDBFeature with static JSON config (#645)
This commit removes the references to the GraalVM SDK and also removes
the AlloyDBFeature class. Instead, the commit uses JSON to configure the
native image.

This resolves a few problems:
- The code wasn't consuming the latest GraalVM SDK from the parent POM
- The presence of newer GraalVM breaks tests on Java 8.
- The native-image META-INF directory wasn't matching
  groupId/artifactId for the Java Connector.

Now, with this commit, there is no need to reference GraalVM SDK at all.
And there is no conflict with Java 8 tests. In addition, there's no need to
keep the native-image.properties file either.

Note that we're using reflect-config.json, the format for GraalVM for JDK 21 [1].
The file name and format changes to reachability-metadata.json for JDK 24 [2].
For now, it seems other Cloud SDK libraries are using the JDK 21 format, so we'll
stick to that.

[1]: https://www.graalvm.org/jdk21/reference-manual/native-image/metadata/#specifying-metadata-with-json
[2]: https://www.graalvm.org/latest/reference-manual/native-image/metadata/#specifying-metadata-with-json